Adding UI Components for Adding Food
1. Updated the MainWindow fxml and code behind class to support the folllowing components.
A. Created a label and textfield for entering the name for a new food.
B. Created a label and combo box for entering the type for a new food.
C. Added a button to the gui for submitting information for a new food.
D. Added a listview to enable the user to select a food from the pantry.