Connect UI to Business Logic for Adding Food
1. Updated the MainWindow class to implement the method for adding a food item.
	I did this by setting the types for ComboBox and ListView.
2. I then set up the comboBox by adding the given types of food from the instructions,
	so when adding a food item the food object is created from using the name from the textfield and type from the combobox.
3. I then implemented the addFood method to be able to add food to the listview.
A. I did this by first getting the food name and type from the textField and combobox.
B. Then I added a try and catch to handle the possibility of an IllegalArgumentException, by alerting the user with a show and wait errorPopup.
C. If the user passes through the try block, it will then add the food name and quantity to the pantry.