Implementing UI for Updating Quantity
1. Updated the MainWindow fxml and codebehind class to support the following:
A. Setting the quantity, added a label, textfield, and a button for setting the quantity of food.
B. Incrementing the quantity, added a button for the user, which I will later add a method to be able to click and increase by one.
C. Decrementing the quantity, added a button for the user, which I will later add a method to be able to click and decrease by one.