Business Logic for Updating Quantity of Food
1. Updated the Food class to support updating quantity of food in the pantry.
2. I did this by implementing a method that allows setting the quantity.
A. This includes passing a value through a parameter to get the quantity value.
B. The value must be greater than or equal to zero as you cant have a negative number of food in the pantry.
C. Also Implemented Unit testing to test the setQuantity is following the precondition and working.
3. Added a method that allows incrementing the quantity.
A. This method increments the quantity by one each time it is called.
B. Implemented Unit testing for this method, to test each call of the method increased the quantity by one.
4. Added a method that allows decrementing the quantity.
A. This method decrements the quantity by one each time it is called.
B. However, this is only true if the quantity is not equal to zero as you cant have a negative number of food in pantry.
C. Implemented Unit testing for this method, to ensure the precondition is working and the method itself is decreasing by one each time it is called.