Connect Business Logic and UI to Update Quantity
1. Implemented methods in the Mainwindow class for setting, incrementing, and decrementing the quantity.
2. Setting the quantity should read the quantity from the textfield and then set the quantity for the food currently selected in the listview.
A. Should check if user is selecting something in the listview and if not should have a show and wait error popup.
B. Should then set the quantity to the food object and refresh the pantry list to display the updated listview.
3. Incrementing the quantity should increment the quantity for the food currently selected in the listview.
A. Should check if the user is selecting something in the listview and if not should have a show and wait error popup.
B. If the user does have something selected it should increase the quantity by one and then refresh the pantry list to display the updated listview.
4. Decrementing the quantity should decrement the quantity for the food currently selected in the listview.
A. Should check if user is selecting something in the listview and if not should have a show and wait error popup.
B. If the user does have something selected it should decrease the quantity by one and then refresh the pantry list to display the updated listview.
C. If the user tries to decrease quantity, by going past zero it should have a show and wait error popup.