Implementation of Sava Bill Data
1. In the Bill Persistence Manager class I implemented the functionality of the saveBillData method.
A. Checked for null and then implemented a try block.
B. In the try block I have a for each loop to go through and write each bill item name and amount seperated by a comma.
2. Implemented the saveBillData in the MainWindow.
A. Did a try block for saving the bill and a catch block to catch an IOException and alert the user with a show and wait if it cant save data to file.