Implementation of Load Bill Data
1. Implemented the method loadBillData in the BillPersistenceManager class.
A. I did this by using a scanner to read through the file.
B. In the try block I first set the server name by using the first line of the file.
C. Then inside the for loop I started the lineNumber at 2 as the first line was already read.
D. While inside the for loop I get the nextLine and strip and split the string into two parts.
E. The two parts are the name and the amount of the item.
2. Added and Implemented a button to load the bill data in the MainWindow.
A. In the loadBillData method in the MainWindow I used a try block to catch the FileNotFoundException and IOException