Implementing Multiple Formats
1. Made a abstract class
A. stripped down save and load
2. Made a CSVBillPersistenceManager class
A. Extends to Abstract class to be able to override the save and load methods.
B. Ensured its comma seperated.
C. Added a toString to return "CSV" so it can say which type it is using.
3. Made a TSVBillPersistenceManager
A. Extends to Abstract class to be able to override the save and load methods.
B. Changed the save and load methods from CSV to TSV, to ensure its tab seperated.
C. Added a toString to return "TSV" so it can say which type it is using.