Implementing Multiple Formats in MainWindow
1. Added a ComboBox to the gui
A. Did this so the user can choose to save the bill in either csv or tsv formats.
2. Changed the savaBillData to get the value of the combo box and save it in that format.
3. Updated the Initialize
A. added the CSV and TSV to the combo box to be able to choose between the two.
B. Also implemented it to load the bill data from either format depending on what is saved