Sorting Bill Items by Cost in MainWindow
1. Added a combo box to the gui to choose between ascending and descending.
A. In the initialize I set the values to be acsending and descending so the user can choose.
2. Added a change order method in mainwindow to sort and update receipt.
A. I noticed my sort wasn't working so I had to add a sort method in the Bill class.
B. Then in the MainWindow I implemented this change to sort the values and then update the receipt.