Added Test Cases To Comparators
1. Made new test packages to organize the test cases by class.
A. Added a testCompare class that tested IngredientNameComparator.
B. Tested when ingredients were null when being compared.
C. Then tested the output integers for each scenario when names are equal, first ingredient already in right spot, and when the ingredient needs to go to the first spot.
2. I then created a new TestToString class to test the toString() in the IngredientNameComparator.
A. Tested the output when toString() is called which results in "Name".
3. Added a testCompare class that tested IngredientTypeComparator.
A. Tested when ingredients were null when being compared.
B. Tested the output integers for each scenario when types are equal, first ingredient already in right spot, and when the ingredient needs to go to the first spot.
4. Created a new TestToString class to test the toString() in the IngredientTypeComparator.
A. Tested the output when toString() is called which results in "Type".
5. Fixed compareTo methods.
A. I changed my compareTo methods in both classes to compareToIgnoreCase, as I noticed a bug when manually testing the gui that didn't sort the ingredients properly.