Updated Business Logic to Add Recipe
1. Implemented a class to store and manage the information for a recipe.
A. Created a Recipe class that stores and manages the information for a recipe.
B. Then wrote a new test case to test the constructor when name is either empty or null, ingredients is empty or null, and when a valid recipe is created.
2. Implemented a utility class to manage converting recipes to String.
A. Created a utility class that converts a recipe to a string.
B. Then wrote a new test case to test the recipeToString method when recipe is null, one ingredient, and two ingredients.
3. Implemented a class to write a recipe to a file.
A. Created a RecipeToFile class that supports writing a recipe to file.
B. Then wrote a new test case to test concerns, while testing I noticed a bug that would happen with my test cases,
which resulted in me writing a @BeforeEach to create a new empty file. This is so when I test, it uses a new file instead of one that has recipe's already in it.
C. I tested when there is a null recipe, one recipe, multiple recipes, and when theres duplicate recipes which would result in an IllegalStateException.