Connected GUI to Business Logic to Add Recipe
1. Implemented MainWindow to support adding a selected ingredient to the ListView for a recipe when the add ingredient recipe button is pressed.
A. Got the ingredient that is currently selected and checked if it was null, if it wasnt it got added to recipe, however if it was null a show and wait error would be triggered.
2. Implemented Mainwindow to create a Recipe object with the name and list of ingredients provided in the TextField and ingredients ListView.
A. Got the name of the recipe and list of ingredients then in a try block I created a new Recipe.
B. I then had multiple catch blocks to catch any exceptions that may get thrown and handled properly by triggering a show and wait errorPopup.