Updated Business Logic to Display Recipes
1.Added a static method that converts a list of recipes to a String, which includes a blank line between each recipe.
A. Then tested the method by checking when the list of recipes are null, empty, one recipe, and multiple.
2. Added a class to load recipes from a file named LoadRecipesFromFile.
A. Added a method that returns a list of recipes loaded from a file, if the file wasnt found or was empty it returned an empy list.
B. This method required a lot of stripping and storing values which I then tested.
C. I tested if there was no recipe in file, one recipe, and two recipes.
3. I then added a method that returns a list of recipes with a specific ingredient included from a file.
A. This method was a lot trickier as I went about it multiple different ways before finally finding a way that worked.
I used the previous method that was just implemented and loaded the recipe list from it.
By doing so I was able to already have the preconditions in their, where if the file was not found or was empty then it returned a empty list.
B. I then tested this method by testing no recipe in file, no recipe in file has the specific ingredient, one recipe in file has specific ingredient,
one recipe in file has specific with multiple ingredients in recipe, and two recipes in file that contain the specific ingredient.