Implemented Input Validation
1.Added a helper method to set up the listener for validation
A. Saw that only one validation is needed as there is one text box that the user will be entering text in.
B. Added the addlistener in the method with a regular expression to make sure the new value matches it.
C. Added a errorText label for the user when typing the minimum length.