Stored List of Passwords
1. Updated the GUI to have a listview instead of a textarea.
A. Did this by first getting rid of the textarea in scene builder and adding a listview instead.
B. I then got rid of any text property for the output and any methods that contained this.
C. I added a ListProperty in the viewmodel and updated the generatePassword method to add the password to the front of the list.
2. Updated test cases to follow the new method behavior
A. Edited the test cases in GeneratePassword view model to check the list property instead of the textproperty.
B. Since we are now storing the passwords in a listview I decided to add a test case that worked with adding multiple valid inputs,
which then stored them in the list.