Loading and Saving Tasks
1. Implemented a class TaskManagerPersistenceManager which has methods to save and load TaskManagerData from specified files.
2. The first method in this class is a static method which saves task manager data.
A. This method takes in a TaskManager and a String of the dataFile name, which both have precondition checks for null and empty.
B. This static method writes the title in the first line then goes to the next line and writes the description and continue.
3. The second method in this class is a static method which loads task manager data.
A. This method takes in a String of a dataFile name, which has the precondition checks for null and empty.
B. This method goes through the file and loads in the tasks to the task manager.