Updated GUI to Support Saving a Set of Tasks
1. Implemented appropriate additions to the viewmodel and bindings for File > Save Tasks menu item,
to save display a file picker then save the current set of tasks to the file selected.
A. Gave a FXID to the Menu Item Save. Made a private helper method for both save and load to clean up the initialize.
B. In MainWindow I also implemented this and made sure to catch any exceptions that can be thrown.
2. In ViewModel I then called this save method to the original method from TaskManagerPersistenceManager.
3. Made sure if the user does not select a file, then the system doesnt perform any action.
4. Made sure if the user selects a file which cant be written over, then the system should display an error stating the file can't be overwritten.
A. First to do this, I did research on the FileWriter and came across the method "canWrite()" which tests whether the application can modify the file.
B. I then used this implementation in the TaskManagerPersistenceManager to throw a SecurityException if the file cant be written in.
C. Then implemented this in MainWindow by having an alert box display to the user if they came across this problem.
5. Then tested the SaveTaskManager from ViewModel, which test null and empty data file, when tasks are in task manager, and when the file is locked.