Updated GUI to Display Add Task Window
1. Updated the MainWindow to display the AddTaskWindow as a modal dialogue when the + button is pressed.
A. In main added a new TASK_WINDOW_TITLE name and a TASK_GUI_RESOURCE to use in MainWindow.
B. This allowed me to set up the button to open another window.
C. This refrenced my new class AddTaskWindow which has a method that sets up binding for view model.
D. Then put this all together in an event handler using the add task button to open the new window.