Updated GUI to Add Task
1.Implemented appropriate properties and methods in the view model to support adding a task given the title and description.
A. Added a String property for task title and task description.
B. Then added a addTask method which got the title and description properties and checked preconditions and then created a new Task.
C. Then added the new task to the task manager.
2. Updated AddTaskWindow to bind onto the appropriate properties and methods from view model.
A. Did this by creating FX ids for the textfield and textarea for the the title and description.
B. I then binded the Task Title and Description and the Title and Description text properties.
C. Then I put the addTask button in an event handler which added the task and then closed the AddTaskWindow when completed.