Updated GUI to Remove Task
1. Implemented appropriate methods in the view model to support removing the currently selected Task from the Task Manager.
A. Added a ObjectProperty in the ViewModel to hold the value of the selectedTask.
B. Added a remove task method to remove the currently selected task in the list view.
C. The method checks the currently selected task and makes sure its not null then removes the task.
2. Implemented appropriate bindings in MainWindow to support removing the currently selected Task when the button is pressed.
A. In the MainWindow the remove task event handler to handle the button calls the method to remove the task in ViewModel.
B. If nothing is selected, an Alert pop up appears asking the user to select a task to remove.